First-Push
Basic functionality being developed - developed utlility functions used
in conjunction with the playerClasses. Currently have implemented the
Player class - not fully finished with it. Aim to use inheritance from
the player class to both a HumanPlayer class and a AIPlayer class. This
should allow for both types of players to be easier to create.