More Classes/Utility updates
Added a new class, Game, with it's own file. Game will be the base class
for various versions of the game, and will handle the core looping,
score tallying, running the players in their order, and determining a
winner. Game is an unfinished class, with two functions that cannot be
tested until it is completed.