Readying game for PVP
Modified game class to have human versus human players.
Fixed a bug in the winner? method - now properly gets winner.
Created a human class which inherits from players.
Created a run method for humans - still need to add reading in a choice
for the human, as well as a way to use cards
Modified the standard player name - who doesn't like the name Bob?!
Created a few methods and functions for printing info to the players.
Updated the rules - forgot the all-important 9-win situation. Fun is now
optional.