Making progress on the game run method
Setup the game run method, also made noted down the code to making a
game instance.
Setup a read in the human run method - need to create a loop for
validation.
Fixed a text bug in print-player-choices.
Created a way to set player names at once - will use at the start of
each pvp game.
Added (make-shuffled-deck) to the end of utilities so it automatically
generates the *deck*