Working on run method
Currently has a few fatal bugs - don't try to run unless you can kill
the program!
Now setup-game at the start of PVP.
Now has a loop for looping through each round. Need to set up with
nested loop.
Created a function for quitting - need to properly check it isn't
garbage.
Created a function for checking if both players are standing - need to
check it isn't garbage.
Fixed a bug in the winner determining process.
Added code in the player run so that when a player chooses to choose a
card, they are given the appropriate prompt.
Fixed a formatting issue in printing the score.