AIPlayer progress
Worked heavily on AIP and the various stuff needed to make it function
properly.
Created AI-Deep-Copy - does as expected. Makes a deep copy of the passed
AI player, and returns the deep copy.
Started on the calculate-successors for the AIP - nearly finished with
it. Need to make a successor for standing.
Created a copy--card utility that copies the card given - it's a deep
copy function.
Cleaned up and removed various TODO's.